% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDatabase.R
\name{createDatabase}
\alias{createDatabase}
\title{Create a Database for Tagging}
\usage{
createDatabase(dbname = getDatabaseName(), mapping, tags, debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{mapping}{an optional list value that specifies a mapping from numerical tag
codes to textual descriptions.  If not provided, a scheme common with CTD
data is used.  (See \sQuote{Examples}.)}

\item{tags}{an optional list that specifies the column names and types for
the tag table.  If not provided, a scheme suitable for CTD data is used.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Create a Database for Tagging
}
\examples{
\dontrun{
library(ocetag)
mapping <- list("good data" = 1, "bad data" = 4)
tags <- list(
    "file" = "TEXT",
    "level" = "INT",
    "tag" = "INT",
    "analyst" = "TEXT",
    "analysisTime" = "TIMESTAMP"
)
unlink("test.db")
db <- createDatabase("test.db", mapping = mapping, tags = tags, debug = 2)
system("echo .dump|sqlite3 test.db")
}

}
\author{
Dan Kelley
}
