% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdtag.R
\name{ctdtag}
\alias{ctdtag}
\title{Run an Shiny App for Tagging CTD Features}
\usage{
ctdtag(
  file = "d201211_0048.cnv",
  prefix = "~/ctdtag",
  plotHeight = 500,
  debug = 0
)
}
\arguments{
\item{file}{character value naming a file to tag. NOTE: this
argument might go away in the future, or at least become optional,
if a directory-searching facility is added.}

\item{prefix}{character value for the start of the name of the database file.}

\item{plotHeight}{numeric value for the height of the plot, in pixels.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
The tags are stored in a SQLite database, for ease of processing in R or other
software systems.  The analyst's name and the time of analysis is stored along
with each time stamp, to facilitate combining the judgements made by multiple
analysts.  A SQLite database is used for this storage of tags.
}
\details{
Instructions are provided with the Help button (or by typing \verb{?}). Tagging
information is stored in a sqlite3 database file, by default named
\code{ctdtag_USERNAME.db}, where \code{USERNAME} is the login name of the analyst.
If this file does not exist, it is created; otherwise, the existing
tags (for the file undergoing analysis) are displayed on the plots,
as a starting point.
}
\author{
Dan Kelley
}
