% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dmsg}
\alias{dmsg}
\title{Optionally print a debugging message}
\usage{
dmsg(debug, ...)
}
\arguments{
\item{debug}{integer value indicating whether to print the message. If this
is 0 or negative, then nothing is printed. If it exceeds 0, the contents of
\code{...} are printed.}

\item{\ldots}{material to be printed.}
}
\description{
Optionally print a debugging message
}
\author{
Dan Kelley
}
