% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabaseName.R
\name{getDatabaseName}
\alias{getDatabaseName}
\title{Get name of database}
\usage{
getDatabaseName(dbprefix = "~/ocetag")
}
\arguments{
\item{dbprefix}{character value used as the start of the name.  The default
value places the database in the user's top level.}
}
\value{
\code{\link[=getDatabaseName]{getDatabaseName()}} returns a character string holding the full
pathname of the database.
}
\description{
Get name of database
}
\author{
Dan Kelley
}
