% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTagMapping.R
\name{getTagMapping}
\alias{getTagMapping}
\title{Get Mapping of Tag Value to Tag Meaning}
\usage{
getTagMapping(dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\value{
\code{\link[=getTagMapping]{getTagMapping()}} returns a data frame of tag value:meaning pairs.
}
\description{
A utility function, used by \code{\link[=ctdTag]{ctdTag()}} to get textual descriptions
of the meanings of the numerical tags.
}
\author{
Dan Kelley
}
