% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTags}
\alias{getTags}
\title{Get Tag Mapping from the Database}
\usage{
getTags(file = NULL, dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{file}{character value naming the CTD file.  Only items relating
to this file are returned.}

\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\value{
\code{\link[=getTagMapping]{getTagMapping()}} returns a data frame holding the \code{code} and
\code{meaning} of each possible tag.
}
\description{
Get Tag Mapping from the Database
}
\author{
Dan Kelley
}
