% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getUserName}
\alias{getUserName}
\title{Get user name}
\usage{
getUserName()

getUserName()
}
\value{
\code{\link[=getUserName]{getUserName()}} returns a character value naming the user, i.e.
holding the user's login name.

\code{\link[=getUserName]{getUserName()}} returns a character value that is the user's
login name on this computer.
}
\description{
Get user name

Get user name
}
\author{
Dan Kelley
}
