% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pluralize}
\alias{pluralize}
\title{Pluralize an item}
\usage{
pluralize(n = 1, singular = "item", plural = NULL)
}
\arguments{
\item{n}{integer giving the number of instances of the item.}

\item{singular}{character value giving the name of the item.}

\item{plural}{optional character value giving the plural name.
If this is not supplied, it is constructed by appending the letter
"s" to \code{singular}.  This lets the output be e.g. "3 geese"
instead of "3 gooses".}
}
\description{
This is useful for saying e.g. "0 apples", "1 apple", or "10 apples".
}
\author{
Dan Kelley
}
