% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatabase.R
\name{readDatabase}
\alias{readDatabase}
\title{Read a tag database into an R object}
\usage{
readDatabase(dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Examine the named database, and return a list containing all
of its data.  This may be helpful to users who are interested
in the internal structure, but who are not cognizant of the R
interface to SQL.
}
\author{
Dan Kelley
}
