% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{removeTag}
\alias{removeTag}
\title{Remove a tag}
\usage{
removeTag(file = NULL, level = NULL, dbname = NULL, debug = 0)
}
\arguments{
\item{file}{character value naming the CTD file.  Only items relating
to this file are returned.}

\item{level}{integer specifying the sequence value of the data point that was tagged.}

\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Remove a tag
}
\author{
Dan Kelley
}
