% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTag.R
\name{saveTag}
\alias{saveTag}
\title{Save a tag}
\usage{
saveTag(
  file = NULL,
  level = NULL,
  scan = NULL,
  tag = NULL,
  analyst = NULL,
  dbname = NULL,
  debug = 0
)
}
\arguments{
\item{file}{character value naming the CTD file.}

\item{level}{integer specifying the sequence value of the tagged data point.}

\item{scan}{integer specifying the scan number of the tagged data point.}

\item{tag}{numerical value of the tag.}

\item{analyst}{character value naming the analyst who is running the app.
The default is to query the operating system to find the person's user-name.}

\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Save a tag
}
\author{
Dan Kelley
}
