% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDatabase.R
\name{updateDatabase}
\alias{updateDatabase}
\title{Update a database of tags}
\usage{
updateDatabase(dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Examine the named database, and update it if required (see
\sQuote{Details}). Report an error if the file does not
exist, or if the file does exist, but it is not an SQLite file
containing tables named \code{version}, \code{tagMapping} and \code{tags}.
}
\details{
The following is a history of changes to the database schema.
\itemize{
\item \strong{2024-04-20:} two changes were made to the database setup.
First, the \code{version} element of the \code{version} table was changed
from 1 to 2.  Second, the \code{level} element of the \code{tags} table was
renamed as \code{index}.  Both changes are automatically handled by
\code{useDatabase()}, which alters old files to the new format. All
other functions of the package use the new format, so the system is
backwards-compatible.
}
}
\author{
Dan Kelley
}
