% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{createDatabase}
\alias{createDatabase}
\title{Create a tagging of database}
\usage{
createDatabase(dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Create a tagging of database
}
\author{
Dan Kelley
}
