% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDatabase.R
\name{createDatabase}
\alias{createDatabase}
\title{Create a Database for Tagging}
\usage{
createDatabase(dbname = getDatabaseName(), mapping, tags, debug = 0)
}
\arguments{
\item{dbname}{character value naming the database.}

\item{mapping}{a named vector (or a list) that specifies a
mapping from textual descriptions to numerical codes; see
\sQuote{Examples}.}

\item{tags}{an optional named vector (or a list) that specifies additional column names and
types for the tag table.  The default set is
\code{list(file = "TEXT", index = "INT", tag = "INT", analyst = "TEXT", analysisTime = "TIMESTAMP")}
which is likely to be suitable for CTD data, with tags referenced to the depth
at the stated index (starting with index 1, at the surface).
See \sQuote{Example} for how pressure may be added.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Create a Database for Tagging
}
\examples{
# This example uses a temporary file and unlinks it later;
# obviously, this is not what would be done in practice, but
# it is necessary for the example to pass CRAN checks.
library(ocetag)
# Use a simple mapping
mapping <- c("good data" = 1, "bad data" = 4)
# Add pressure to the 'tags' database
tags <- c("pressure" = "FLOAT")
dbname <- tempfile() # do not do this in practice
# Finally, create the database.
createDatabase(dbname, mapping = mapping, tags = tags)
unlink(dbname) # do not do this in practice

}
\author{
Dan Kelley
}
