% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{ctdTag}
\alias{ctdTag}
\title{Run an R-shiny app to facilitating tagging features on CTD profile data.}
\usage{
ctdTag(file = "d201211_0048.cnv")
}
\arguments{
\item{file}{character value naming a file to tag. NOTE: this
argument might go away in the future, or at least become optional,
if a directory-searching facility is added.}
}
\description{
Run an R-shiny app to facilitating tagging features on CTD profile data.
}
\author{
Dan Kelley
}
