% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTags}
\alias{getTags}
\title{Get Tag Mapping from the Database}
\usage{
getTags(file = NULL, dbname = getDatabaseName(), debug = 0)
}
\arguments{
\item{file}{character value naming the CTD file.  Only items relating
to this file are returned.}

\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\value{
\link{getTags} returns a data frame holding the tags that
have been saved for a particular data file.
}
\description{
Get Tag Mapping from the Database
}
\examples{
# This example uses a temporary file and unlinks it later;
# obviously, this is not what would be done in practice, but
# it is necessary for the example to pass CRAN checks.
dbname <- tempfile() # do not do this in practice
useDatabase(dbname)
saveTag("ctd.cnv", index = 1, tag = 2, analyst = "Dan Kelley", dbname = dbname)
tags <- getTags("ctd.cnv", dbname = dbname)
print(tags)
unlink(dbname) # do not do this in practice

}
\author{
Dan Kelley
}
