% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTag.R
\name{saveTag}
\alias{saveTag}
\title{Save a tag}
\usage{
saveTag(
  file,
  index,
  ...,
  tag,
  analyst = "anonymous",
  dbname = getDatabaseName(),
  debug = 0
)
}
\arguments{
\item{file}{character value naming the CTD file.}

\item{index}{integer specifying the sequence value of the tagged
data point.}

\item{\dots}{optional extra columns to be stored in the database.
These must have been specified in the \code{\link[=useDatabase]{useDatabase()}} call
that created the database named in \code{dbname}.}

\item{tag}{numerical value of the tag.}

\item{analyst}{character value naming the analyst who is running the app.
The default is to query the operating system to find the person's user-name.}

\item{dbname}{character value naming the database.}

\item{debug}{integer value indicating a debugging level.  Use 0 for silent
operation, or any integer exceeding 1 to show some output during the action
of this function.}
}
\description{
Save a tag
}
\examples{
# This example uses a temporary file and unlinks it later;
# obviously, this is not what would be done in practice, but
# it is necessary for the example to pass CRAN checks.
dbname <- tempfile() # do not do this in practice
useDatabase(dbname)
saveTag("ctd.cnv", index = 1, tag = 2, analyst = "Dan Kelley", dbname = dbname)
tags <- getTags("ctd.cnv", dbname = dbname)
print(tags)
unlink(dbname) # do not do this in practice

}
\author{
Dan Kelley
}
